requirements.txt
pandas
numpy
requests
scikit-learn
xgboost
joblib
matplotlib
seaborn
streamlit
python-dotenv
jupyter
notebook
